+++ caldav_compat_tests/__init__.py
@@ -1,0 +1,22 @@
+# This file is part of Radicale - CalDAV and CardDAV server
+# Copyright © 2026-2026 Arnav S <172543153+theofficialtruck@users.noreply.github.com>
+#
+# This library is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This library is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with Radicale.  If not, see <http://www.gnu.org/licenses/>.
+
+"""
+CalDAV compatibility tests for Radicale, based on the ``caldav`` client
+library and the ``caldav-server-tester`` compatibility-checking tool.
+
+See README.md in this directory for details.
+"""